requirements.txt
load-dotenv==0.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
